Add simple_pick_place ros2 package
This ros2 package is designed to be built and run from a workspace with all of MoveIt2 source code and tutorials also built in it (as tutorials have guided). As such, program may not run in a different environment. This program must also be run after starting URSim, launching the ur_robot_driver and ur_moveit_config launch files (with launch_rviz:=true).